Tarefa1

Description

INTRODUCTION

In this Module we intended to create an incomplete starter Maze generated by random numbers, surround it with walls and fill it with tunnels and put the Ghosts' house at the center. At this part, the segment that consumed more time to make, was to place the tunnels and the Ghosts' house.

STRATEGY

We started by creating functions that would provide us with random numbers, then we converted them to the type Pieces. Next, we created lists made by Pieces, we changed them to Corridors' and then finally to a Maze (because a Maze, by default, is a listing of Pieces' lists - Corridors'). The next step was to set the Maze into a box. In other words, we're surrounding it with walls. To accomplish this, we had to create a function that generates Corridors to fill the upper and lower bounds of the Maze. Then we preceded and proceeded each corridor with walls. Now, with every field filled, we need to create tunnels. To do so, we had to locate the middle of the Maze's height and replace the walls (before and after the labyrinth in the coordinates found) with with the Empty Pieces. Ultimately, to build the Ghosts' house, we created the default house structure, for both conditions: whether the Maze's length is either even or odd (if it is even, then the Ghost's house length is 10; if it isn't, its length is 11) and placed it into the middle of the Maze.

CONCLUSION

As far as we are concerned, this task wasn't very difficult. Despite that , we believe to have done a great job, considering that all goals were accomplished and that the task at hand was easy to understand.

Generating random numbers

maze :: [[Piece]]

Maze

generateRandoms :: Int -> Int -> [Int]

Given a seed returns a list of n integer randomly generated; it takes the first n elements from an infinite series of random numbers between 0-99

nrRandoms :: Int -> Int

Given a seed returns an integer randomly generated

subList :: Int -> [a] -> [[a]]

Converts a list into a list with size n

Creating pieces and printing them

convertPiece :: Int -> Piece

Converts an integer number into a Piece

convertCorridor :: [Int] -> Corridor

Converts a integers' list into a Corridor

printCorridor :: Corridor -> String

Converts a Corridor to a String

convertMaze :: [[Int]] -> Maze

Converts a list of integers' lists into a Maze

printMazes :: Maze -> String

Converts a Maze to a String

Putting the Maze into a box

wallCorridor :: Int -> Corridor

Creates a Corridor of Walls

wallSides :: Maze -> Maze

Creates Walls in each Corridor

wallBox :: Maze -> Maze

Creates Walls all around the Maze

getMiddle :: Int -> Int

Get the middle of an Int

replaceinList :: Int -> Piece -> Corridor -> Corridor

Exchanges an index for a Piece

replaceinCorridorAB :: Maze -> Int -> Piece -> Maze

Get the Walls after and before the Maze

putTunnel :: Maze -> Maze

Put Tunnels in the Maze

Ghosts' house

buildHouse :: Maze -> Maze

Creates the Ghosts' House with specific measures: if the Mazes's length is even, the Ghosts' House's length is 10, otherwise is 11

concatghosts :: Maze -> Maze -> Maze

Put the ghosts' pieces together

replaceinCorridor :: Corridor -> Corridor -> Corridor

Exchanges pieces for the Ghosts' House does that for the first Corridor of the Ghosts' House

getNotMiddle :: Maze -> Maze -> Maze

Exchanges pieces for the Ghosts' House does that for the rest of the Corridors of the Ghosts' House

ghostshouse :: Maze -> Maze

Builds the Ghosts' House

Final functions

genMazeR :: Int -> Int -> Int -> Maze

Generates the Maze randomly

genMaze :: Int -> Int -> Int -> Maze

Generates a valid Maze

generateMaze :: Maze -> IO ()

Transforms Maze string into IO ()

Examples to test

maze1R :: Maze

Example 1 genMazeR

maze2R :: Maze

Example 2 genMazeR

maze1w :: Maze

Example 1 wallBox

maze2w :: Maze

Example 2 wallBox

maze1p :: Maze

Example 1 putTunnel

maze2p :: Maze

Example 2 putTunnel

maze1 :: Maze

Example 1

maze2 :: Maze

Example 2