Tarefa2

Description

INTRODUCTION

In this task we wanted to set the function Play, where given the current state of the game and a movement made by a player, it determines the subsequent state of the game. Of all the tasks in the first part of the project, this was undoubtedly the most challenging.

STRATEGY

Here, we had to consider several variants: the Players (that can be Ghost's or a Pacman's), their states (in other words, their identification, coordinates, velocity, orientation, points and number of remaining lifes) and the actions (for example, if the Player is a Pacman and there's a small consumable next to him, it should move for that position, eat the food and win one more point). In this task, we had to define each subject and it's sub-divisions precisely so that we could evaluate each move's possibilities. However, in this module We only set the Plays for the Pacman, as requested by the teaching team.

CONCLUSION

This task was one of the most crucial ones, since it was this step that allowed us to start reaching the end result. We were successful in thinking and getting all the cases for a Play. Although we are happy with the results, we believe that our code could be cleaner and simplified.

All players or pacman management

idplayer :: Int -> [Player] -> Player

Creates the identification (id) of a Player

absorvePiece :: Maze -> Coords -> Maze

Do the Pacman's path

nextPiece :: Maze -> Coords -> Orientation -> Piece

Returns the piece before some coords

pieceNext :: Maze -> Coords -> Orientation -> Bool

Checks if there's a piece next

spookyscary :: [Player] -> Coords -> Bool

Verify if there is a ghost in certain coords

checkghost :: Player -> GhostMode

See if a Ghost is dead or alive

chooseGhost :: [Player] -> Coords -> Player

Select the ghost from the list of Players

newplayer :: Player -> Orientation -> Player

Adds a new Player into the Maze in certain orientation

subPlayer :: Player -> [Player] -> [Player]

Substitues Players

coordstunnel :: Coords -> Maze -> Coords

Pass through the empty tunel to the other side of the Maze

auxo :: Orientation -> Coords -> Coords

Transform the coords with an eye on the orientation

addscore :: Int -> Player -> Player

Add integers into score of Players

turnsmega :: Player -> Player

Turns into mega

lives :: Player -> Player

Takes lives from Players, when 0 equals Dying

liv :: [Player] -> [Player]

Takes lives from Pacman's, when hitting Ghost's

Ghosts management

changeGhostOrientation :: PlayerState -> PlayerState

Changes Ghost's orientation when they're death

allghostsdead :: [Player] -> [Player]

Turns all ghosts to dead

allghostsalive :: [Player] -> [Player]

Turns all ghosts to alive

aliveghost :: Player -> Player

Changes the state of a ghost to alive

pacmanNormal :: Player -> Player

Auxiliar of turnsNormal

turnsNormal :: Int -> [Player] -> [Player]

Turn Pacman into Normal again

Final Functions

validPlay :: Play -> State -> State

Do a Play if it's valid

play :: Play -> State -> State

Do all the Plays with all conditions