Tarefa3

Description

INTRODUCTION

This module aims to convert a Maze into a Instruction's list to give an easy reading. The difficult part of this task was finding an error on the optimizeInstructions function. It was, briefly, executing on an infinite cycle. However, with some patience and research we got rid of that infinite cycle, making it work properly.

STRATEGY

An Instruction is a Tuplo's list, where the first value is an Int and the second one is a Piece. After reading these instructions, the interpreter should return the Maze (since the sum of all the Int should result in the Maze's width and the length of that list must correspond to the Maze's height). Our goal was to create a good compression function, to minimize the number of Instructions to be given towards the interpreter. With this in mind, we firstly create some funtions for convertions. We wanted to get Duplos of pieces and their quantity from single pieces and get Instructions not optimized from a Maze. Once the convertions were concluded, we examine if a certain Instruction is valid and ultimately create the patterns (vertical and horizontal) with the optimized Instructions.

CONCLUSION

Much like the first task, this one wasn't very difficult either. We have successfully obtained a function that given a Maze turns it into as few Instructions lists as possible to give to the interpreter.

Defining Instructions

pieceToDuplo :: Piece -> Duplo

From a normal Piece, goes to a Duplo

corridorToDuplos :: Corridor -> [Duplo]

Turns a normal Corridor into a Duplo's list

mazeToDuplos :: Maze -> [[Duplo]]

Turns a normal Corridor into a list of Duplo's lists

corridorToInstructions :: Corridor -> Instruction

Turns Duplo's lists into a Instruction not optimized

mazeToInstructions :: Maze -> Instructions

After getting a normal maze transform it into Instructions not optimized

Verifying if the Instructions are valid

doubleSum :: [Duplo] -> Int

Aux that adds the quantity of total pieces

validInstruction :: Corridor -> Instruction -> Bool

Valid instructions for Corridor

validInstructions :: Maze -> Instructions -> Bool

Valid Instructions for Maze

Building standards

optimizedInstructions :: [Duplo] -> Instruction

Turns the accumulator into a corridor of Instruction

duploaccum :: [Duplo] -> [Duplo]

Accumulates Instructions' Corridor

optimizedInstructionsm :: [[Duplo]] -> Instructions

Turns the accumulator into a Maze of Instructions

repeatInstructions :: Int -> Instructions -> Instructions

Repeat Instructions in every case

rInstructions :: Int -> Instructions -> Instructions

Put the Repeat in on case

poselem :: Instructions -> Int -> Instruction

Auxiliar function where we get the position of one Instruction in a list of Instructions' lists

accumR :: Int -> Instructions -> Instructions

Accumulates the compare fact between Corridors

generalcompact :: Instructions -> Instructions

Loop of accumR

Final function

compactMaze :: Maze -> Instructions

Transform a Maze into the Instructions method