Tarefa4

Description

INTRODUCTION

In this task, we intended to calculate the time's effect on the game's state. To do so, we needed to define the function passTime, which given a step and the actual game's state, it updates the game's state after a repetition. The most difficult part was making the Ghosts move.

STRATEGY

This task's goals are to define time and its influences on the game. In which we wanted to make the Players move after the time set so we could register a new movement. To accomplish that, we have to consider some restrictions, such as: time cant stop, and by that we mean that the

(continua porque eu não sei a estrategia usada)

CONLUSION

Despite the difficulties we felt with regard to the players' movements, especially the Ghosts, we believe that we did a great job (porque nao sei o que...)

Documentation

defaultDelayTime :: Integer

250 ms is the default delay time

comparesteps :: Int -> Int -> Bool

Compare Step's before and after Mega

walkonthengbh :: Player -> State -> State

Make the Players' move around

updateTimeMega :: Double -> Player -> Player

Update time Mega

reduceTM :: [Player] -> [Player]

Auxiliar of reduceTimeMega

reduceTimeMega :: State -> State

Reduces time Mega

FINAL FUNCTION

auxpT :: Int -> [Player] -> State -> State

Auxiliar of final function

passTime :: Int -> State -> State

Final Function