Tarefa5

Description

INTRODUCTION

This task's main objective is to implement moves for the Ghosts as well.

STRATEGY

The most effective way to obtain moves from the Ghosts, was to add play function conditions. However, this is insubstantial. Their moves must react to Pacman's moves accordingly, since the movement of the Ghosts is dictated by Pacman's state. Given his Normal state, they must chase him; oppositely, the Ghosts must run away when Pacman changes his state to the Mega state. Hence is determined that when the Ghosts are in the Alive mode, they are also in chaseMode, otherwise (Dead mode), in scatterMode. To achieve effective chasing dexterity, the Ghosts must locate the best, shortest path to reach Pacman. To do so, we implemented a function to calculate that path. Another required restriction, is to know how will Pacman react upon hitting a wall: he will have to rotate his directional movement to the right (in other words: if he's moving to the right, then, upon hitting a wall, he will now change to a downwards trajectory. Same applies to all other directions).

CONCLUSION

We believe that this task was successful, since we did all of the game's demanded restrictions in an efficient manner. As far as we are concerned, this was the best way to conclude the task.

WHEN THE GHOSTS ARE ALIVE (chasemode)

distance :: Coords -> Coords -> Float

Calculates the shortest distance between two players

sdistance :: Orientation -> Coords -> Coords -> Orientation

Auxiliar of chasemode

chasemode :: State -> Int -> Play

Final function when Ghosts' are Alive

WHEN THE GHOSTS ARE DEATH (scattermode)

clockwisedead :: Orientation -> Orientation

Rotation clockwise when death

scattermode :: State -> Int -> Play

Final function when Ghosts' are Dead

FINAL FUNCTION

list :: [a] -> Int -> a

Same as the function (!!)

ghostPlay :: State -> [Play]

Final Function

ghostsPlay :: Int -> [Play] -> Play

Plays of different Ghosts'