Tarefa6

Description

INTRODUCTION

This task's main objective in to implement an bot that plays the Pacman based on the current state of the game and that it performs a singular command using the type Play

STRATEGY

Since we need to cover all aspects and all plays that Pacman's can do it was a very long and careful task for us. For the start we decided to divide this task in three, something look a like others tasks and for that we started when Pacman is Mega. After that we created a radar with a raio of 1 and looking for Food Fig, not finding that, it would go the direction more away from the closest Ghost that had Food Little, if there were no Food Little we would just run from Ghosts till it find Food Big or Little.

CONCLUSION

This was one of the hardest quests but for sure it was the one that brang us the most amount of errors and concerns about beucase it was the joint of the most important tasks.

WHEN MEGA

goaway :: State -> Coords

Get's closer Ghost from Pacman's

sdistancePm :: Maze -> Orientation -> Coords -> Coords -> Orientation

Run for Ghosts' when Mega

chasemodeP :: State -> Int -> Play

Final function when Pacman's is Mega

RADAR OF PACMAN

sdistancePM :: Maze -> Orientation -> Coords -> Coords -> Orientation

Run from Ghosts' when not Mega

radar :: Maze -> PlayerState -> Piece -> Coords

See Pieces' around Pacman's in a radar of 1

radarn :: Maze -> PlayerState -> Piece -> Coords

See Pieces' around Pacman's in a radar of 1 in all orientation's

radarR :: Maze -> Player -> Piece -> Coords

Restricted radar when close to limits

WHEN THERE ARE FOOD BIG

pacmanshow :: Maze -> Player -> Piece -> Coords -> Orientation

If there are Food Big's around Pacman's radar goes that way

searchFB :: State -> Piece -> Play

Search for a Food Big's in radar and do a play

checksFB :: Maze -> Coords -> Bool

Checks if there are Food Big's in radar

WHEN THERE ARE NO FOOD BIG

notpacmanshow :: Maze -> Player -> Piece -> Coords -> Orientation

If there are Food Little's around Pacman's radar goes that way

searchFL :: State -> Piece -> Play

Search for a Food Little's in radar and do a play

checksFL :: Maze -> Coords -> Bool

Checks if there are Food Little's in radar

FINAL FUNCTION

botFood :: Int -> State -> Play

Final Function without when Pacman's is not Mega

bot :: Int -> State -> Maybe Play

Final function

maybeToPlay :: Maybe Play -> Play

Turns bot plays into actual plays

auxpTB :: Int -> [Player] -> State -> State

Make the bot play

passTimeBot :: Int -> State -> State

Final Function for the bot