TypesandDatas

Description

Documentation

data State

state

Constructors

State

Fields

maze :: Maze

playersState :: [Player]

level :: Int

type Maze = [Corridor]

list of corridors

type Corridor = [Piece]

list of pieces

data Piece

types of pieces

Constructors

Food FoodType

PacPlayer Player

Empty

Wall

data Player

player type: pacman or ghost

Constructors

Pacman PacState

Ghost GhoState

data Orientation

orientation : left, right, up, down or null

Constructors

L

R

U

D

Null

data PacState

pacman state

Constructors

PacState

Fields

pacState :: PlayerState

timeMega :: Double

openClosed :: Mouth

pacmanMode :: PacMode

data GhoState

ghost state

Constructors

GhoState

Fields

ghostState :: PlayerState

ghostMode :: GhostMode

data AllGhosts

ghosts names

Constructors

AllGhosts

Fields

clyde :: GhoState

blinky :: GhoState

inky :: GhoState

pinky :: GhoState

data Manager

main manager

Constructors

Manager

Fields

state :: State

pid :: Int

step :: Int

before :: Integer

delta :: Integer

delay :: Integer

type Duplo = (Int, Piece)

duplo

type Coords = (Int, Int)

coords

type PlayerState = (Int, Coords, Double, Orientation, Int, Int)

player state

data Mouth

mouth open or closed

Constructors

Open

Closed

data PacMode

pacman mode

Constructors

Dying

Mega

Normal

data GhostMode

ghost mode

Constructors

Dead

Alive

data FoodType

food type

Constructors

Big

Little

data Color

colors

Constructors

Blue

Green

Purple

Red

Yellow

None

data Play

play type

Constructors

Move Int Orientation

type Instructions = [Instruction]

list on instruction

data Instruction

duplo to instruction

Constructors

Instruct [(Int, Piece)]

Repeat Int

coloredString :: String -> Color -> String

Returns a String with a given color

placePlayersOnMap :: [Player] -> Maze -> Maze

Puts the given players on the Maze

printMaze :: Maze -> String

Plots the Maze in graphical symbols

printPlayerStats :: Player -> String

Shows the Player stats

getPlayerID :: Player -> Int

Get the Player's identity (ID)

getPlayerPoints :: Player -> Int

Get the Player's points

setPlayerCoords :: Player -> Coords -> Player

Set the Player's coordinates

getPlayerCoords :: Player -> Coords

Get the Player's actual coordinates

getPlayerSCoords :: PlayerState -> Coords

Get the Player State's actual coordinates

getPacman :: [Player] -> Player

Get Pacman's out of a list

getGhosts :: [Player] -> [Player]

Get Ghosts' from list of players

getPacmanMode :: Player -> PacMode

Get Pacman's mode (Dying | Mega | Normal)

getGhostMode :: Player -> GhostMode

Get Ghosts' mode (Dead | Alive)

getPlayerState :: Player -> PlayerState

Get the Player's state

setPlayerOrientation :: Player -> Orientation -> Player

Set the Player's orientation

getPlayerOrientation :: Player -> Orientation

Get the Player's actual orientation

getPiece :: Maze -> Coords -> Piece

Get Piece's in the maze

getPieceOrientation :: Piece -> Orientation

Get the Piece's orientation

getPlayerVelocity :: Player -> Double

Get the Player's velocity

replaceElemInMaze :: Coords -> Piece -> Maze -> Maze

Put the given piece on the choosen coordinates, replacing the old one

replaceNElem :: Int -> a -> [a] -> [a]

Replace a specific element of a list with a given element

getTimeMega :: [Player] -> Double

Gets the time when the Pacman is Mega

getMouth :: Player -> Mouth

Get the Player's actual mouth status

changeinMouth :: Player -> Player

Change the Player's mouth status

getStep :: Manager -> Int

Get step

isPacman :: Player -> Bool

See if the Player's is Pacman

getMaze :: State -> Maze

Get Maze's out of a State's

getPlayers :: State -> [Player]

Get Players' out of State's